@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>Mohammad Moattar</Authors>
     <Company>LittleBlocks</Company>
@@ -8,14 +8,14 @@
     <PackageProjectUrl>https://github.com/LittleBlocks/LittleBlocks.Ef</PackageProjectUrl>
     <RepositoryUrl>https://github.com/LittleBlocks/LittleBlocks.Ef</RepositoryUrl>
     <Version>1.0.0</Version>
-    <LangVersion>10</LangVersion>
+    <LangVersion>12</LangVersion>
     <PackageIcon>Logo.png</PackageIcon>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="10.1.1" />
-    <PackageReference Include="LittleBlocks" Version="1.0.1" />
-    <PackageReference Include="NSubstitute" Version="4.2.2" />
-    <PackageReference Include="FluentAssertions" Version="6.8.0" />
+    <PackageReference Include="AutoMapper" Version="13.0.1" />
+    <PackageReference Include="LittleBlocks" Version="2.0.0" />
+    <PackageReference Include="NSubstitute" Version="5.1.0" />
+    <PackageReference Include="FluentAssertions" Version="6.12.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\LittleBlocks.Ef\LittleBlocks.Ef.csproj" />
